Replace pencil mark Set with array
Specific implementation incompatible with Jest - seems like a bug in the
Jest source code "Set.prototype.values called on incompatible receiver"

A few issues opened in the Jest repo in the past but all ignored and
auto-closed after stale. Not own error, as Set.prototype.values is never
manually called in this codebase.
Seems to be an error with how .toEqual recursively checks things, losing
Sets' "this" binding.